Track Chair App Authz
Added new group "track-chairs-admins"

+++ database/code/CustomDatabaseAdmin.php
@@ -21,7 +21,9 @@ class CustomDatabaseAdmin extends DatabaseAdmin
 {
     public function build() {
         // drop index if exists
-        DB::query("ALTER TABLE `Member` DROP INDEX `ExternalUserId`;");
+        $res = DB::query("SHOW KEYS FROM `Member` WHERE Key_name='ExternalUserId'")->first();
+        if($res)
+            DB::query("ALTER TABLE `Member` DROP INDEX `ExternalUserId`;");
 
         parent::build();
         $this->dropObsoleteTables();
